+++ crates/sui-graphql-rpc/tests/examples_validation_tests.rs
@@ -1,10 +1,6 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-use similar::*;
-use std::fs::File;
-use std::io::Read;
-use std::path::PathBuf;
 #[cfg(feature = "pg_integration")]
 mod tests {
     use rand::rngs::StdRng;
@@ -114,35 +110,5 @@ mod tests {
             bad_examples.queries.len(),
             "all examples should fail"
         );
-    }
-}
-
-#[test]
-fn test_generate_markdown() {
-    let mut buf: PathBuf = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
-    buf.push("docs");
-    buf.push("examples.md");
-    let mut out_file: File = File::open(buf).expect("Could not open examples.md");
-
-    // Read the current content of `out_file`
-    let mut current_content = String::new();
-    out_file
-        .read_to_string(&mut current_content)
-        .expect("Could not read examples.md");
-    let new_content: String = sui_graphql_rpc::examples::generate_markdown()
-        .expect("Generating examples markdown failed");
-
-    if current_content != new_content {
-        let mut res = vec![];
-        let diff = TextDiff::from_lines(&current_content, &new_content);
-        for change in diff.iter_all_changes() {
-            let sign = match change.tag() {
-                ChangeTag::Delete => "---",
-                ChangeTag::Insert => "+++",
-                ChangeTag::Equal => "   ",
-            };
-            res.push(format!("{}{}", sign, change));
-        }
-        panic!("Doc examples have changed. Please run `sui-graphql-rpc generate-examples` to update the docs. Diff: {}", res.join(""));
     }
 }
